IotModelFunctionApp/requirements.txt
azure-functions
pandas
pyodbc
scikit-learn
pytz
